Fix a lock contention in Signaller controller
The lock contention was observed under a relatively high amount of PURGE
requests, that Signaller propagates among instances in a Varnish
cluster, and changes in endpoints of a K8s service for Varnish;

The Signaller controller locks the "Signaller" structure when it tries
to get/update one of the fields (endpoints) in that structure.

This commit reduces the time when the "Signaller" structure is locked by
copying the current set of endpoints instead of locking the structure
while the controller sends PURGE requests.