Watch: Drop previous VCL version when reloading.
As stated by @timkante in issue #81, Varnish keeps all
previous VCL versions after reloads. This increases the
memory consumption of Varnish over time and can be deadly
if the configuration gets repeatedly reloaded (because of
autoscaling related behaviors for example).

This commit intends to keep to 1 the list of VCLs in the
history.