Varnish: Change default port from 80 to 8080.
The port 8080 is a "standard" HTTP port for applications running inside
a private infrastructure. As the default HTTP port (80) is a system port
, it is not possible to restrain permissions through the securityContext
and drop rootUser out-of-the-box. This will result in a permission error
while trying to create a socket on port 80.

This commit intends to fix this and make the application of the securityContext
(required in most if not every production clusters) a lot more simple.